+++ plugins/clock_monotonic/clock_monotonic.c
@@ -11,7 +11,7 @@ time_t uwsgi_monotonic_seconds() {
 uint64_t uwsgi_monotonic_microseconds() {
         struct timespec ts;
         clock_gettime(CLOCK_MONOTONIC, &ts);
-	return (ts.tv_sec * 1000000) + (ts.tv_nsec/1000);
+	return ((uint64_t)ts.tv_sec * 1000000) + (ts.tv_nsec/1000);
 }
