+++ plugins/clock_realtime/clock_realtime.c
@@ -11,7 +11,7 @@ time_t uwsgi_realtime_seconds() {
 uint64_t uwsgi_realtime_microseconds() {
         struct timespec ts;
         clock_gettime(CLOCK_REALTIME, &ts);
-	return (ts.tv_sec * 1000000) + (ts.tv_nsec/1000);
+	return ((uint64_t) ts.tv_sec * 1000000) + (ts.tv_nsec/1000);
 }
